+++ Lab/Lab1-1/Program.cs
@@ -4,21 +4,21 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Lab1_1
+namespace Lab1_2
 {
-    class BubbleSort
+    class BubbleSortFunction
     {
         static void Main(string[] args)
         {
-            bool flag = true;
+
 
             // input
             int[] input = { 4, 5, 2, 8, 9, 1, 2, 4, 3, 1 };
 
             // process
-            while (flag)
+            for (int p = 0; p != input.Length - 1; p++)
             {
-                flag = false;
+
                 for (int i = 0; i != input.Length - 1; i++)
                 {
                     if (input[i] < input[i + 1])
@@ -26,10 +26,11 @@ namespace Lab1_1
                         int temp = input[i];
                         input[i] = input[i + 1];
                         input[i + 1] = temp;
-                        flag = true;
+
                     }
                 }
             }
+
 
             // output
             for (int i = 0; i != input.Length; i++)
